add: debug tests
* Big and small tests are added. Big tests check the attachment for the
process and the full script for debugging two processes:
launching the application, installing breakpoints, connecting to another process,
installing the breakpoints on it, yhe transition  between the breakpoints.
* Tests that attach to non-child processes require ptrace_scope to be set to 0,
initially in the system this is not the case.
Therefore, if the system is not configured, then such tests are skipped.
* In .vscode obsolete settings for ruff lsp have been removed
and settings for the bandit have been added.